main and clearance
