+++ code/logManagement.cpp
