--- code/pe+mem/README
@@ -1,2 +1,0 @@
-Para compilar: g++ pe.cpp -o pe
-Para ejecutar: ./pe
