--- state_enum.cpp
@@ -1,12 +1,0 @@
-#ifndef STATE_ENUM_H
-#define STATE_ENUM_H
-
-enum StateEnum {
-    Modified, 
-    Owned, 
-    Exclusive, 
-    Shared, 
-    Invalid
-};
-
-#endif
